auth_api/requirements.txt
alembic==1.11.1
async-timeout==4.0.2
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.4
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
speaklater==1.3
SQLAlchemy==2.0.15
typing_extensions==4.6.2
unicorn==2.0.1.post1
uvicorn==0.22.0
Werkzeug==2.3.4
WTForms==3.0.1
zope.event==4.6
zope.interface==6.0
opentelemetry-api==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-exporter-jaeger==1.18.0
